requirements.txt
pandas>=2.0.0
openai>=1.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-generativeai>=0.8.0
tableau-api-lib>=0.1.36
asyncio-throttle>=1.0.0
